Rename package to si3exit. Bump to v2.0.0
This is done because on Manjaro it conflicts with the base i3exit.
Only way to get around it is to ignore updates to the package from
upstream. But now that I have my own Arch Linux repository, that would
mean it would ignore my own i3exit upstream package as well. And since
I can't just delete the base i3exit package from Manjaro because
apparently it's depended by settings packages, it's simply easier to
rename mine and move on.